Mod some names to public scope :bulb:
I've modified some of the names on this repo, to allow modification on some methods.

For example, I've intended to fork just the bdd file to another package I've created. For this, I've needed to make a fork on this project and make some variables and functions public to allow some changes, while maitaining the previous functionality.