fix(mspec): Fix dep cycle on mspec test by moving :bug:
I've fixed the dep cycle, by moving the mspec tests to a proper folder.
It's a temporary solution, althought I don't think there will be a
better one now.
To fix the dep cycle without isolating the tests on its own package,
would require a lot of refactoring, on all packages. But this could be
made on a different PR.